Move Auth to Separate Package
