+++ internal/auth/auth_test.go
@@ -1,4 +1,4 @@
-package util
+package auth
 
 import (
 	"errors"
